pyproject.toml
[tool.poetry]
name = "data-reliability-playbook"
version = "0.1.0"
description = "Code examples and demos for my NDSML Summit 2025 talk: \"The Data Reliability Ceiling: A Technical Playbook for AI-Ready Observability.\""
authors = ["Eugene Klyuchnikov <eugene.klyuchnikov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"  # Specifies Python 3.11.x is required

# Core dependencies for all plays
pyyaml = "^6.0.1"
jsonschema = {extras = ["format"], version = "^4.25.1"}
pandas = "^2.2.2"
whylogs = {extras = ["viz"], version = "^1.6.4"}
ipykernel = "^6.29.4" # For running jupyter notebooks in the venv
openai = "^1.35.7"   # For the bonus demo

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"